+++ schemas/service.py
@@ -1,4 +1,5 @@
 from datetime import datetime
+from decimal import Decimal
 from typing import Optional, List
 
 from pydantic import BaseModel, Field
@@ -26,7 +27,7 @@ class ServiceCreate(BaseModel):
     name: str = Field(min_length=2, max_length=120)
     description: Optional[str] = None
     duration_minutes: int = Field(gt=0)
-    price: float = Field(ge=0)
+    price: Decimal = Field(ge=0, decimal_places=2)
     points_reward: int = Field(default=0, ge=0)
 
 
@@ -35,7 +36,7 @@ class ServiceUpdate(BaseModel):
     name: Optional[str] = Field(None, min_length=2, max_length=120)
     description: Optional[str] = None
     duration_minutes: Optional[int] = Field(None, gt=0)
-    price: Optional[float] = Field(None, ge=0)
+    price: Optional[Decimal] = Field(None, ge=0, decimal_places=2)
     points_reward: Optional[int] = Field(None, ge=0)
     is_active: Optional[bool] = None
 
@@ -47,7 +48,7 @@ class ServiceResponse(BaseModel):
     name: str
     description: Optional[str]
     duration_minutes: int
-    price: float
+    price: Decimal
     points_reward: int
     is_active: bool
     created_at: datetime
